+++ spec/rubocop/cli/cli_auto_gen_config_spec.rb
@@ -441,6 +441,47 @@ RSpec.describe RuboCop::CLI, :isolated_environment do
       end
     end
 
+    context 'when working with a cop who do not support auto-correction' do
+      it 'can generate a todo list' do
+        create_file('example1.rb', <<~RUBY)
+          def fooBar; end
+        RUBY
+        create_file('.rubocop.yml', <<~YAML)
+          # The following cop does not support auto-correction.
+          Naming/MethodName:
+            Enabled: true
+        YAML
+        expect(cli.run(%w[--auto-gen-config])).to eq(0)
+        expect($stderr.string).to eq('')
+        # expect($stdout.string).to include('Created .rubocop_todo.yml.')
+        expect(Dir['.*']).to include('.rubocop_todo.yml')
+        todo_contents = IO.read('.rubocop_todo.yml').lines[8..-1].join
+        expect(todo_contents).to eq(<<~YAML)
+          # Offense count: 1
+          # Configuration parameters: EnforcedStyle, IgnoredPatterns.
+          # SupportedStyles: snake_case, camelCase
+          Naming/MethodName:
+            Exclude:
+              - 'example1.rb'
+
+          # Offense count: 1
+          # Cop supports --auto-correct.
+          # Configuration parameters: EnforcedStyle.
+          # SupportedStyles: always, always_true, never
+          Style/FrozenStringLiteralComment:
+            Exclude:
+              - 'example1.rb'
+        YAML
+        expect(IO.read('.rubocop.yml')).to eq(<<~YAML)
+          inherit_from: .rubocop_todo.yml
+
+          # The following cop does not support auto-correction.
+          Naming/MethodName:
+            Enabled: true
+        YAML
+      end
+    end
+
     context 'when working in a subdirectory' do
       it 'can generate a todo list' do
         create_file('dir/example1.rb', ['$x = 0 ',